requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
email-validator
bcrypt
python-jose
